+++ odoo/js/terminal.js
@@ -1049,8 +1049,21 @@ odoo.define("terminal.Terminal", function (require) {
             if (ev.keyCode === 27) {
                 // Press Escape
                 this.doHide();
-            } else if (ev.ctrlKey && ev.key === "1") {
-                // Press Ctrl + 1
+            } else if (
+                (ev.ctrlKey && ev.key === "1") ||
+                (ev.altKey && ev.key && ev.key.toLowerCase() === "t")
+            ) {
+                // Press Alt + t (and mark as deprecated CTRL + 1)
+                if (
+                    !this._advDeprecatedShortcut &&
+                    ev.ctrlKey &&
+                    ev.key === "1"
+                ) {
+                    this.screen.print(
+                        "<strong style='color:orange'>** INFO: The terminal has been opened using a keyboard shortcut that will no longer be supported in the next release because it collides with browser shortcuts. Please use 'ALT + T' to open the terminal.</strong>"
+                    );
+                    this._advDeprecatedShortcut = true;
+                }
                 ev.preventDefault();
                 this.doToggle();
             }